==> src/02_Feature_Engineering.ipynb <==
import pandas as pd

# National Medal Counts

games = pd.read_pickle('../dat/games.pkl')
games.head()

games['Summer']     = games['Season'] == 'Summer'
games['Female']     = games['Gender'] == 'F'
games['Gold']       = games['Medal']  == 'Gold'
games['Silver']     = games['Medal']  == 'Silver'
games['Bronze']     = games['Medal']  == 'Bronze'
games['Home Field'] = games['NOC']    == games['Host Country']
games.head()

# total sport and event participation by games and country
medals = pd.DataFrame(games.groupby(['Games','NOC','Region'])['Athlete','Sport','Event'].nunique())
medals.head()

# total females and gold, silver and bronze medals by games and country
df     = pd.DataFrame(games.groupby(['Games','NOC','Region','Athlete'])['Female'].mean())
df     = df.groupby(['Games','NOC','Region']).sum()
medals = medals.merge(df, left_index=True, right_index=True)
medals.head()

# total medal count by games and country
df     = pd.DataFrame(games.groupby(['Games','NOC','Region','Event'])['Medal'].nunique())
df     = df.groupby(['Games','NOC','Region']).sum()
medals = medals.merge(df, left_index=True, right_index=True)
medals.head()

# total medal count by games and country
df     = games.groupby(['Games','NOC','Region','Event'])['Gold','Silver','Bronze'].sum()
df     = df.clip(upper=1)
df     = df.groupby(['Games','NOC','Region']).sum()
medals = medals.merge(df, left_index=True, right_index=True)
medals.head()

# season, year and home-field advantage by games and country
df     = pd.DataFrame(games.groupby(['Games','NOC','Region'])['Summer','Year','Home Field'].mean())
medals = medals.merge(df, left_index=True, right_index=True)
medals.head()

medals = medals.reset_index()
medals.head()

medals['Female']     = medals['Female'].astype('int64')
medals['Gold']       = medals['Gold'].astype('int64')
medals['Silver']     = medals['Silver'].astype('int64')
medals['Bronze']     = medals['Bronze'].astype('int64')
medals['Summer']     = medals['Summer'].astype('int64')
medals['Home Field'] = medals['Home Field'].astype('int64')
medals.head()

medals = medals.rename(columns={'Athlete':    'Athletes',
                                'Medal':      'Medals',
                                'Female':     'Females',
                                'Gold':       'Golds',
                                'Silver':     'Silvers',
                                'Bronze':     'Bronzes',
                                'Sport':      'Sports',
                                'Event':      'Events',
                                'Home Field': 'Host'})
medals.head()

medals = medals[['Year',
                 'Summer',
                 'Games',
                 'Host',
                 'NOC',
                 'Region',
                 'Athletes',
                 'Females',
                 'Sports',
                 'Events',
                 'Medals',
                 'Golds',
                 'Silvers',
                 'Bronzes']]
medals.head()

medals['Athletes per Event'] = (medals['Athletes'] / medals['Events']).round(3)
medals.head()

# save dataframe
medals.to_pickle('../dat/medals.pkl')
