==> src/04_Linear_Regression.ipynb <==
from   sklearn.model_selection import train_test_split
from   sklearn.preprocessing   import MinMaxScaler
from   sklearn.linear_model    import LinearRegression
from   sklearn.metrics         import mean_squared_error, r2_score
import pandas                  as     pd
import numpy                   as     np
import seaborn                 as     sns
import matplotlib.pyplot       as     plt

# Dummy Encoding

medals = pd.read_pickle('../dat/medals.pkl')
medals.head()

# create dummy variables
model = pd.get_dummies(data=medals, columns=['NOC'])
model.head()

# Correlation

# Train Test Split

y = pd.DataFrame(model['Medals'])
y.head()

X = model.drop(['Medals','Games','Golds','Silvers','Bronzes','Region'], axis=1)
X.head()

# Scaling
scaler = MinMaxScaler()
scaler.fit(X)
scaler.transform(X)

# Linear Regression

# Create linear regression object
regr = LinearRegression()

# Train the model using the training sets
regr.fit(X[['Year','Summer','Host','Athletes','Females','Sports','Events']], y)

# Make predictions using the testing set
y_pred = pd.DataFrame(regr.predict(X), columns=['Prediction'])
y_pred['Prediction'] = y_pred['Prediction'].astype('int64')
y_pred.head()

y['Prediction'] = y_pred['Prediction']
y['NOC']        = medals['NOC']
y['Error']      = y['Medals'] - y['Prediction']
y               = y.sort_values(by='Medals', ascending=False)
y

# Validation on Test Set

X.columns

# The coefficients
features = pd.DataFrame(regr.coef_.T, columns=['Coefficients'])
features['Feature'] = X.columns.T
features = features.sort_values(by='Coefficients', ascending=False)
features

regr.intercept_

regr.score(X,y)

# Explained variance score: 1 is perfect prediction
print('Variance score: %.2f' % r2_score(y, y_pred))

# The mean squared error
print("Mean squared error: %.2f" % np.sqrt(mean_squared_error(y, y_pred)))

# Validation with 2016 Rio Olympics

# Prediction of 2020 Tokyo Olympics

# Plot outputs
plt.figure(figsize = (10,6))
sns.set_style("whitegrid")
plt.scatter(X['Events'], y, c='navy', alpha=0.6)
plt.plot(X['Events'], y_pred, color='red', linewidth=3)
plt.show()
