==> src/05_Model_Validation.ipynb <==
import plotly
import plotly.plotly     as     py
import plotly.graph_objs as     go
import pandas            as     pd
from   config            import plotly_id, plotly_key

plotly.tools.set_credentials_file(username=plotly_id, api_key=plotly_key)

medals = pd.read_pickle('../dat/medals.pkl')
medals.head()

rio_2016_medals = medals[medals['Games'] == '2016 Summer']
rio_2016_medals.head()

hover_text = []
color_range = []
for index, row in rio_2016_medals.iterrows():
    hover_text.append(('Games: {games}<br>'+
                       'Host: {host}<br>'+
                       'Nation: {noc}<br>'+
                       'Region: {region}<br>'+
                       'Athletes: {athletes}<br>'+
                       'Females: {females}<br>'+
                       'Sports: {sports}<br>'+
                       'Events: {events}<br>'+
                       'Medals: {medals}<br>'+
                       'Golds: {golds}<br>'+
                       'Silvers: {silvers}<br>'+
                       'Bronzes: {bronzes}').format(games    = row['Games'],
                                                    host     = row['Host'],
                                                    noc      = row['NOC'],
                                                    region   = row['Region'],
                                                    athletes = row['Athletes'],
                                                    females  = row['Females'],
                                                    sports   = row['Sports'],
                                                    events   = row['Events'],
                                                    medals   = row['Medals'],
                                                    golds    = row['Golds'],
                                                    silvers  = row['Silvers'],
                                                    bronzes  = row['Bronzes']))
    color_range.append(max(row['Athletes'],100))

trace0 = go.Scatter(
    x      = rio_2016_medals['Events'],
    y      = rio_2016_medals['Medals'],
    text   = hover_text,
    mode   = 'markers',
    marker = dict(
#        size          = rio_2016_medals['Athletes'],
        color         = color_range,
        showscale     = True,
        reversescale  = True,
        colorscale    = 'RdBu',
        colorbar      = dict(
                title = 'Total<br>Athletes'
            )
    )
)

data      = [trace0]
layout    = go.Layout(
    title = 'Medal Table of the Rio 2016 Summer Olympic Games',
    xaxis = dict(
        title         = 'Total Number of Events Participated',
        gridcolor     = 'rgb(255, 255, 255)',
        zerolinewidth = 1,
        ticklen       = 5,
        gridwidth     = 2,
    ),
    yaxis = dict(
        title         = 'Total Number of Medals Won',
        gridcolor     = 'rgb(255, 255, 255)',
        zerolinewidth = 1,
        ticklen       = 5,
        gridwidth     = 2,
    ),
    paper_bgcolor = 'rgb(243, 243, 243)',
    plot_bgcolor  = 'rgb(243, 243, 243)',
)

fig = go.Figure(data = data, layout = layout)
py.iplot(fig, filename = 'rio-2016-olympic-medals')
